moved spinner into loaders folder
